fix launch queue issues

+++ atup-core/src/main/java/org/feuyeux/jaxrs2/atup/core/constant/AtupVariable.java
@@ -2,7 +2,7 @@ package org.feuyeux.jaxrs2.atup.core.constant;
 
 public interface AtupVariable {
     long DETECT_INTERVAL = 10;//10 seconds
-    long DETECT_CONNECT_TIMEOUT = 20;//20 seconds
+    int DETECT_CONNECT_TIMEOUT = 20;//20 seconds
     long TESTING_TIMEOUT = 10000;//10 seconds
     long ASYNC_LAUNCH_TEST_TIMEOUT = 300000;//5 minutes
 }
